chore: move login and register to modules - remove getElementById from navbar - finished removing indented html
