+++ src/app/modules/about-us/pages/about-us/about-us.component.ts
@@ -1,5 +1,5 @@
-import {Component, OnInit, ViewEncapsulation} from '@angular/core';
-import {TranslateService} from '@ngx-translate/core';
+import {Component, OnInit} from '@angular/core';
+import { AppConfigService } from '../../../../core/services/configuration/app-config.service';
 
 @Component({
   selector: 'app-about-us',
@@ -9,9 +9,11 @@ import {TranslateService} from '@ngx-translate/core';
 
 export class AboutUsComponent implements OnInit {
 
-  constructor() {
+  constructor(private appConfigService: AppConfigService) {
     scroll(0, 0);
   }
 
-  ngOnInit(): void {}
+  ngOnInit(): void {
+    this.appConfigService.setTitle('ABOUTUS.PAGETITLE');
+  }
 }
