+++ src/app/modules/authentication/register/register.component.css
@@ -3,7 +3,7 @@ html {
   height: auto;
 }
 
-.login-container {
+.main-container {
   background: rgb(255, 255, 255);
   background: linear-gradient(to bottom, #f0f0f0 0%, white 100%);
   background-size: cover;
