+++ src/app/modules/authentication/register/register.component.spec.ts
