+++ src/app/modules/noticias/pages/noticias/noticias.component.css
@@ -1,16 +1,12 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@100;400&display=swap');
 
-html {
-    background-size: cover;
-}
 
-body {
-    padding: 10%;
+.main-container {
     background: rgb(255, 255, 255);
     background: linear-gradient(to bottom, #f0f0f0 0%, white 100%);
     background-size: cover;
     font-family: 'Roboto', Arial, Helvetica, sans-serif;
-    padding-top: 160px;
+    padding: 160px 10% 10%;
 }
 
 .topbottomLine {
