+++ src/app/shared/components/news-card/news-card.component.css
@@ -6,6 +6,7 @@
     margin: 5px 0;
     display: flex;
     flex-direction: column;
+    width: 100%;
 }
 .news-card-img {
     align-self: center;
