+++ src/app/shared/components/news-card/news-card.component.html
@@ -1,11 +1,11 @@
 <a [routerLink]="['/noticias/' + noticia.reference]">
-<mat-card class="news-card">
-    <img appResize [ngWidth]="400" class="news-card-img" mat-card-image
-    ngSrc="{{noticia.imageUrl}}" alt="Foto de Noticia"
-    >
-    <mat-card-content class="news-card-content">
-        <mat-card-title class="clamp">{{ noticia.title }}</mat-card-title>
-        <mat-card-subtitle class="clamp align-end">{{ noticia.author }}</mat-card-subtitle>
-    </mat-card-content>
-</mat-card>
+  <mat-card class="news-card">
+      <img appResize [ngWidth]="400" class="news-card-img" mat-card-image
+      ngSrc="{{noticia.imageUrl}}" alt="Foto de Noticia"
+      >
+      <mat-card-content class="news-card-content">
+          <mat-card-title class="clamp">{{ noticia.title }}</mat-card-title>
+          <mat-card-subtitle class="clamp align-end">{{ noticia.author }}</mat-card-subtitle>
+      </mat-card-content>
+  </mat-card>
 </a>
