+++ src/assets/i18n/es.json
@@ -159,7 +159,8 @@
     },
     "ABOUTUS": {
       "TITLE": "Sobre nosotros",
-      "LEARNMORE": "Ver más"
+      "LEARNMORE": "Ver más",
+      "PAGETITLE": "ITBA IEEE | About us"
     },
     "IEEENEWS": {
       "TITLE": "Noticias",
@@ -239,6 +240,7 @@
     }
   },
   "ABOUTUS": {
+    "PAGETITLE": "ITBA IEEE | About us",
     "US": {
       "TITLE": "Quiénes somos",
       "TEXT": "IEEE-ITBA es una organización estudiantil sin fines de lucro que representa al IEEE en el Instituto Tecnológico de Buenos Aires."
@@ -293,6 +295,8 @@
     "LEVEL-TEXT": "La dificultad de los problemas es variada para asegurar que los menos experimentados también puedan participar."
   },
   "NAVBAR": {
+    "LOGIN": "Iniciar sesión",
+    "WELCOME": "Bienvenido ",
     "ABOUTUS": "Acerca de",
     "CONTACT": "Contacto",
     "EVENTS": "Eventos",
@@ -625,6 +629,7 @@
   },
   "ASIMOVCUP": {
     "TITLE": "Asimov Cup",
+    "PAGETITLE": "ITBA IEEE | Asimov CUP",
     "ENROLLBUTTON": "¡Anotate ahora!",
     "DOWNLOAD_RULEBOOK": "Descargar reglamento",
     "CATEGORIES": {
