+++ src/index.html
@@ -4,7 +4,7 @@
   <meta charset="utf-8">
   <title>ITBA IEEE Stud. Branch</title>
   <base href="/">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1">
   <link rel="icon" type="image/x-icon" href="assets/image/general-icons/IEEEITBA%20Icono.svg">
   <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500&display=swap" rel="stylesheet">
   <link href="https://fonts.googleapis.com/css2?family=Inconsolata:wght@300;400;500&display=swap" rel="stylesheet">
